refactor module structure and class names
