+++ src/main/java/com/mds/journal_app/mapper/JournalMapper.java
@@ -1,7 +1,7 @@
 package com.mds.journal_app.mapper;
 
 
-import com.mds.journal_app.pojo.Journal;
+import com.mds.journal_app.dao.Journal;
 import com.mds.journal_app.pojo.JournalResponse;
 import org.mapstruct.Mapper;
