+++ src/main/java/com/mds/journal_app/pojo/JournalEntryRequest.java
@@ -1,14 +1,12 @@
 package com.mds.journal_app.pojo;
 
+
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
 @Data
-@Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class PostJournalRequest {
-    String title;
-    String description;
+public class JournalEntryRequest {
+    String textContent;
 }
