+++ src/main/java/com/mds/journal_app/pojo/JournalRequest.java
@@ -5,13 +5,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
-import java.time.Instant;
-
 @Data
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class PostJournalResponse {
+public class JournalRequest {
     String title;
     String description;
-    Instant dateCreated;
 }
