--- src/main/java/com/mds/journal_app/pojo/PostJournalEntryRequest.java
@@ -1,12 +1,0 @@
-package com.mds.journal_app.pojo;
-
-
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class PostJournalEntryRequest {
-    String textContent;
-}
