+++ src/test/java/com/mds/journal_app/JournalControllerTest.java
@@ -1,0 +1,4 @@
+package com.mds.journal_app;
+
+public class JournalControllerTest {
+}
