--- src/test/java/com/mds/journal_app/JournalResponseAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mds.journal_app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JournalResponseAppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
